requirements.txt
aws-cdk-lib==2.128.0
constructs>=10.0.0,<11.0.0
boto3==1.34.49
PyYAML==6.0.1
